displaycolor: Skip init for external displays
Recent change to enable displaycolor pipeline initialization for external
displays has caused a memory-on-boot regression. This code path is not
fully ready and not enabled yet anyway, so skip external display when
initializing displaycolor for now. When we switch DisplayPort to use
DECON1 for output, or when displaycolor implements a lazy pipeline init,
we will re-enable this.

Bug: 308255833
Bug: 288608645
Test: presubmit
Change-Id: Ie1f0e48423c91d1b36a56799d6eb01b60a2e0709